rename : renamed( Ocilo ), change logRegister to register
