+++ .github/workflows/publish.yml
@@ -1,4 +1,4 @@
-name: Publish vmlog to PyPI
+name: Publish Ocilo to PyPI
 on:
   release:
     types: [published]
