+++ src/Ocilo/ScopeResolver.py
