+++ src/Ocilo/__init__.py
@@ -3,22 +3,22 @@ import threading
 from importlib.metadata import PackageNotFoundError, version
 
 from . import vmlog_engine  # Import eagerly so a missing C extension fails at import time.
-from ._core import _VMlog
+from ._core import _Oscilo
 
 try:
-    __version__ = version("vmlog")
+    __version__ = version("Ocilo")
 except PackageNotFoundError:
     __version__ = "unknown"
 
-__all__ = ["logRegister"]
+__all__ = ["register"]
 
 # The package owns exactly one monitor instance; it is created lazily on first
-# use so that importing vmlog alone has no side effects (no tracing, no atexit hook).
+# use so that importing Ocilo alone has no side effects (no tracing, no atexit hook).
 _instance = None
 _instance_lock = threading.Lock()
 
 
-def logRegister(varName):
+def register(varName):
     global _instance
 
     if _instance is None:
@@ -26,7 +26,7 @@ def logRegister(varName):
         # instances, which would reintroduce the multi-instance bug (issue #29).
         with _instance_lock:
             if _instance is None:
-                _instance = _VMlog()
+                _instance = _Oscilo()
 
     # Pass the caller's frame explicitly because the instance sits one call deeper.
     _instance.register(varName, sys._getframe(1))
