+++ src/Ocilo/_core.py
@@ -7,8 +7,8 @@ from .ScopeResolver import ScopeResolver
 from .TraceDispatcher import TraceDispatcher
 
 
-class _VMlog:
-    def __init__(self, fileName="vmlog.jsonl"):
+class _Oscilo:
+    def __init__(self, fileName="Ocilo.jsonl"):
         self.fileName = fileName
         self.buffer = HistoryBuffer()
         self.dispatcher = TraceDispatcher(self.buffer)
