+++ src/Ocilo/vmlog_engine.c
