+++ tests/test_vml_components.py
@@ -4,10 +4,10 @@ import sys
 import tempfile
 import unittest
 
-from vmlog._core import _VMlog
-from vmlog.FileWriter import FileWriter
-from vmlog.HistoryBuffer import HistoryBuffer
-from vmlog.ScopeResolver import ScopeResolver
+from Ocilo._core import _Oscilo
+from Ocilo.FileWriter import FileWriter
+from Ocilo.HistoryBuffer import HistoryBuffer
+from Ocilo.ScopeResolver import ScopeResolver
 
 LOCAL = 0
 GLOBAL = 1
@@ -116,7 +116,7 @@ class TestVMlogComponents(unittest.TestCase):
             filename = os.path.join(temp_dir, "deleted_event.jsonl")
 
             target = [1, 2]
-            monitor = _VMlog(filename)
+            monitor = _Oscilo(filename)
             monitor.register("target")
 
             target.append(3)
@@ -141,7 +141,7 @@ class TestVMlogComponents(unittest.TestCase):
             filename = os.path.join(temp_dir, "idempotent_save.jsonl")
 
             target = ["start"]
-            monitor = _VMlog(filename)
+            monitor = _Oscilo(filename)
             monitor.register("target")
 
             target.append("changed")
