+++ tests/test_vml_edge_cases.py
@@ -3,7 +3,7 @@ import os
 import tempfile
 import unittest
 
-from vmlog._core import _VMlog
+from Ocilo._core import _Oscilo
 
 GLOBAL_COLLISION_VALUE = "global-value"
 INIT_EVENT = "init"
@@ -33,7 +33,7 @@ class TestVMlogEdgeCases(unittest.TestCase):
             filename = os.path.join(temp_dir, "no_duplicate.jsonl")
 
             target = [1, 2, 3]
-            monitor = _VMlog(filename)
+            monitor = _Oscilo(filename)
             monitor.register("target")
 
             # Run several traced lines without changing the tracked value.
@@ -61,7 +61,7 @@ class TestVMlogEdgeCases(unittest.TestCase):
             filename = os.path.join(temp_dir, "nested_mutable.jsonl")
 
             target = {"users": [{"name": "Alice", "score": 10}]}
-            monitor = _VMlog(filename)
+            monitor = _Oscilo(filename)
             monitor.register("target")
 
             target["users"][0]["score"] = 20
@@ -83,7 +83,7 @@ class TestVMlogEdgeCases(unittest.TestCase):
 
             def run_local_scope():
                 target_name_collision = "local-value"
-                monitor = _VMlog(filename)
+                monitor = _Oscilo(filename)
                 monitor.register("target_name_collision")
 
                 target_name_collision = "local-updated"
@@ -108,7 +108,7 @@ class TestVMlogEdgeCases(unittest.TestCase):
             filename = os.path.join(temp_dir, "stop_tracking.jsonl")
 
             target = [1]
-            monitor = _VMlog(filename)
+            monitor = _Oscilo(filename)
             monitor.register("target")
 
             target.append(2)
@@ -141,7 +141,7 @@ class TestVMlogEdgeCases(unittest.TestCase):
             filename = os.path.join(temp_dir, "unicode.jsonl")
 
             target = {"message": "안녕하세요", "status": "준비"}
-            monitor = _VMlog(filename)
+            monitor = _Oscilo(filename)
             monitor.register("target")
 
             target["status"] = "완료"
@@ -163,7 +163,7 @@ class TestVMlogEdgeCases(unittest.TestCase):
             number = 1.5
             values = (1, 2)
 
-            monitor = _VMlog(filename)
+            monitor = _Oscilo(filename)
             monitor.register("flag")
             monitor.register("number")
             monitor.register("values")
